+++ .env.production
@@ -5,5 +5,8 @@ NEXT_PUBLIC_ENV=PROD
 ### server api
 NEXT_PUBLIC_API_URL=https://book-piseo.com
 
+### Naver Book API
+NEXT_PUBLIC_NAVER_BOOK_API_URL=https://openapi.naver.com
+
 #### next auth
 NEXT_PUBLIC_AUTH_URL=http://localhost:3000
